% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creat_age.R
\name{create_age}
\alias{create_age}
\title{Introduction to R packages; creating a Fake People}
\usage{
create_age()
}
\value{
Return a value for your age
}
\description{
Introduction to R packages; creating a Fake People
}
