% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name.R
\name{give_name}
\alias{give_name}
\title{Title}
\usage{
give_name(sex = "female")
}
\arguments{
\item{sex}{input is either "male" or "female"}
}
\value{
returns one name
}
\description{
Title
}
\examples{
give_name()
give_name(sex = "male")
}
