% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemand-package.R
\docType{package}
\name{iemand-package}
\alias{iemand}
\alias{iemand-package}
\title{iemand: Test Package for Cape Town}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andrew Clark \email{agcur@rogers.com}

}
\keyword{internal}
