% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise_nicely.R
\name{praise_nicely}
\alias{praise_nicely}
\title{Praise folk!}
\usage{
praise_nicely(name)
}
\arguments{
\item{name}{A person's name}
}
\value{
A string of praise in upper case.
}
\description{
Praise folk!
}
\examples{
praise_nicely("Bob")
}
