% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/za_names.R
\docType{data}
\name{sa_names}
\alias{sa_names}
\title{Names given to babies born in 2016 in South Africa}
\format{A data.frame with columns first_name, number (frequency), and sex.}
\source{
Stats SA
}
\description{
Data about live births collected by Stats SA, see \url{http://www.statssa.gov.za/?page_id=1854&PPN=P0305&SCH=7092}.
The data is distributed with a CC-BY SA license.
}
