% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age.R
\name{create_age}
\alias{create_age}
\title{Create a randomised age}
\usage{
create_age()
}
\value{
age (integer)
}
\description{
Create a randomised age
}
\details{
This is a pointless function for training purposes :)
}
