% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age.R
\name{create_age}
\alias{create_age}
\title{Create an age for someone}
\usage{
create_age()
}
\value{
age (Integer)
}
\description{
Create an age for someone
}
\details{
This is a useless function we created to learn! Gives an age
}
\examples{
none
}
