% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age.R
\name{create_age}
\alias{create_age}
\title{Create an age for someone}
\usage{
create_age()
}
\value{
age (integer)
}
\description{
Create an age for someone
}
\details{
Wee, this is the best function. The best. Using
a uniform distribution capped at 1 and 100 when ages go beyond both,
and are not uniformly distributed.
}
\examples{
create_age()

}
\author{
Constance Neeser
}
