% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name.R
\name{give_name}
\alias{give_name}
\title{function with data}
\usage{
give_name(sex = "female")
}
\arguments{
\item{sex}{female or male}
}
\value{
returns an assigned name given gender (name)
}
\description{
function with data
}
