% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name.R
\name{give_name}
\alias{give_name}
\title{Gives a name depending on sex assigned at birth}
\usage{
give_name(sex = "female")
}
\arguments{
\item{sex}{sex assigned at birth, "male" or "female"}
}
\description{
Gives a name depending on sex assigned at birth
}
\examples{
give_name()
give_name(sex = "male")
}
