% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name.R
\name{give_name}
\alias{give_name}
\title{Create function that gives name}
\usage{
give_name(sex = "female")
}
\arguments{
\item{sex}{gender of name to return}
}
\value{
character name
}
\description{
Samples 1 name
}
\examples{
give_names("female")
give_names("male")

}
