% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name.R
\name{give_name}
\alias{give_name}
\title{Pick a random name based on given sex}
\usage{
give_name(sex = "female")
}
\arguments{
\item{sex}{sex listed for name}
}
\value{
a name from the specified group
}
\description{
Pick a random name based on given sex
}
