% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name.R
\name{give_name}
\alias{give_name}
\title{Gives a name based on assigned gender}
\usage{
give_name(sex = "female")
}
\arguments{
\item{sex}{(character)}
}
\value{
Returns a name
}
\description{
Gives a name based on assigned gender
}
