% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give_name.R
\name{give_name}
\alias{give_name}
\title{Get a random, South African name.}
\usage{
give_name(gender = "female")
}
\arguments{
\item{gender}{The gender of the name you want.}
}
\value{
A string
}
\description{
Get a random, South African name.
}
\examples{
give_name("male")
give_name()
}
