% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemand-package.R
\docType{package}
\name{iemand-package}
\alias{iemand}
\alias{iemand-package}
\title{iemand: What the Package Does (One Line, Title Case)}
\description{
R package workshop
}
\author{
\strong{Maintainer}: Andrew Henning \email{andrew.henning@digitata.com}

}
\keyword{internal}
