% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemand-package.R
\docType{package}
\name{iemand-package}
\alias{iemand}
\alias{iemand-package}
\title{iemand: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
