% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemand-package.R
\docType{package}
\name{iemand-package}
\alias{iemand}
\alias{iemand-package}
\title{iemand: Creates Fake People}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jaymon0703/iemand}
  \item Report bugs at \url{https://github.com/jaymon0703/iemand/issues}
}

}
\author{
\strong{Maintainer}: Jasen Mackie \email{jaymon0703@gmail.com}

}
\keyword{internal}
