% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemand-package.R
\docType{package}
\name{iemand-package}
\alias{iemand}
\alias{iemand-package}
\title{iemand: Iemand}
\description{
Creates fake people.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/L3min/iemand}
  \item Report bugs at \url{https://github.com/L3min/iemand/issues}
}

}
\author{
\strong{Maintainer}: Lara Basson \email{laraebasson@gmail.com}

}
\keyword{internal}
