% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemand-package.R
\docType{package}
\name{iemand-package}
\alias{iemand}
\alias{iemand-package}
\title{iemand: Create Fake People}
\description{
This package creates fake people.
}
\author{
\strong{Maintainer}: Marie Dussault \email{mcdussault@hotmail.co.uk}

}
\keyword{internal}
