% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iemand-package.R
\docType{package}
\name{iemand-package}
\alias{iemand}
\alias{iemand-package}
\title{Create random human data}
\description{
Create data. Nice data. The best of data.
}
\details{
This is more detail. The best detail. Can't get enough detail.
}
\author{
Constance Neeser
}
\keyword{internal}
